Keycard authentication device rework (#2648)
Keycard auth device uses no power
Keycard auth device now can be used to start a countdown for red alert or ship abandonment. The countdown is 3 minutes. If nothing was done during the countdown, it starts an event (red alert or ship abandonment).
The countdown can be cancelled through any KCAD by a crew member with sufficient access. Then countdown stops and no event is called.
The countdown can be proceeded through any KCAD by a crew member with sufficient access if they are not the member with a card that started the countdown. Then the countdown ends and calls event immediately.
Emergency Maint Access is given and revoked without a countdown through any KCAD by a crew member with sufficient access.