Update (#3)
* QOL Door interactions, apc lights, inhand ID door interaction (#2601)

* QOL Door interactions, apc lights

* review fixes

* review fixes

* Adds else ifs to telecomms machine deconstruction steps (#2615)

* Adds returns to telecomms machine deconstruction steps

They were both triggering on a single action. now they won't.

* Now with less bad code

* Missed one

Should be done now

* Fixs Crossbow Crafting and Adds Description Steps (#2631)

Previously Crossbows couldn't be crafted the expected way, with the screwdriver step being doable earlier, and finishing it early.
As well, added to the descriptions to give some idea to folks unfamiliar with it what the next step is to build it.

* Fix robotic limbs descriptions typo. (#2633)

* Allows Printing of 38 Ammo (#2635)

* Fixs Crossbow Crafting and Adds Description Steps

Previously Crossbows couldn't be crafted the expected way, with the screwdriver step being doable earlier, and finishing it early.
As well, added to the descriptions to give some idea to folks unfamiliar with it what the next step is to build it.

* Adds Autolathe Recipes for 38 ammo.

* Adds Autolathe Recipes to Appropriate Disks

* Fix clonepod name (#2636)

"clonpod" -> "clon**e**pod"

* QOL tweaks to access (#2634)

* Attempt to remove loading icon time in rnd machinery (#2627)

* Items in hands now preffered then ears for :l :r (#2637)

* Makes hardhats normal-sized item (#2641)

* Embedding Fixes (#2639)

* Fixed embedding

* thrown embedding

* Adds the ability to revert mech DNA lock (#2646)

* Keycard authentication device rework (#2648)

Keycard auth device uses no power
Keycard auth device now can be used to start a countdown for red alert or ship abandonment. The countdown is 3 minutes. If nothing was done during the countdown, it starts an event (red alert or ship abandonment).
The countdown can be cancelled through any KCAD by a crew member with sufficient access. Then countdown stops and no event is called.
The countdown can be proceeded through any KCAD by a crew member with sufficient access if they are not the member with a card that started the countdown. Then the countdown ends and calls event immediately.
Emergency Maint Access is given and revoked without a countdown through any KCAD by a crew member with sufficient access.

* Fix and Tweaks for chemestry (#2649)

Few improvements in chemestry code mostly style related.

* pref flavour fix (#2643)

* Fixes roundstart issues (#2642)

A small package of fixes to solve problems with starting rounds:

Storyteller start requirements changed. Instead of captain, now any head of staff is required. Someone's still got to be in charge til the captain wakes up, but he can join later.
This should improve captain RP by not forcing someone to hastily take the role

Admin force-start now works correctly and actually starts the round

The storyteller is no longer reset to guide if starting fails